feat: Mirror an image horizontally and vertically
This is a relatively trivial task. It involves swapping upper / left
with lower / right side. Two cases arise: horizontal mirroring and
vertical mirroring. They are handled by maintaining a boolean variable.
The processing should not affect the original image in memory so a new
object is made. Direct assignment does not work as it leads to a
shallow copy of the BufferedImage object. Hence the deepCopy method
is written.