feat: Rotation using math that became necessary
Image needs to be rotated with a user-specified angle in degrees. There
are two slightly shorter methods to do this. They are:
(i) mapping source pixels to destination pixels and
(ii) deriving destination pixels from source pixels.
Method (i) does not produce accurate results in many cases because of
sine and cosine values getting rounded up which leads to a wave-like
pattern in the result. Method (ii) produces accurate but grainy images.
So we build up on top of method (ii)'s accuracy and bring in binary
interpolation.