feat: Use sobel operator to detect edges
An edge is formed where there is a significant difference in the
instensity of pixels. The operator uses two 3×3 kernels which are
convolved with the original image to calculate approximations of the
derivatives – one for horizontal changes, and one for vertical. We take
the magnitude of these found values (vector addition). We do this for
all three layers (R,G,B) of every pixel.