refactor: Avoid * during imports
The import directive should not be burdened with unwanted classes.
When a class does not exist in the given file, the compiler searches
for the class from the list of imports which implies that using * for
imports is increasing the search space for the compiler. Another reason
to avoid * is to prevent naming conflicts if any.