feat: implement gaussian blur using box blur
State of the art algorithm for blur is gaussian blur. But it is
slow. It has a time complexity of O(height * width * intensity^2). We
cannot do that. Then we have a box blur algorithm whose naive time
complexity is O(height * width * intensity^2). Due to central limit
theorem, repeated box blur results into an approximation of gaussian
blur. We can separate the kernel in box blur and apply it horizontally
and vertically. That brings down the time complexity to
O(height * width * intensity). If we pay attention to how the window
slides linearly, we can further optimize it to O(height * width) which
is what has been implemented.