feat: Posterize an image
Image posterizing depends upon reducing the number of distinct pixels
in an image. The rules for pixel reduction can vary. We reduce a pixel
in constant time by keeping only 4 possible output values for each input
value ranging from 0-255. The choice of the values and the number of
values is arbitrary. A possible future addition could be to let the end
user decide this.