feat: Added AES256CTR.java
Individual functions are well-detailed with comments.
AES256 encrypts a single (16 byte) block of data. To encrypt more, we
need a mode of operation. There are several with their pros and cons:
  1. ECB - fast but not secure
  2. CBC - secure but cannot support parallelism
  3. CTR - supports parallelism, secure, no integrity check
  4. GCM - essentially CTR + integrity check
The choice between GCM and CTR was tough. However, the project is
meant to be run locally and not hosted somewhere. And since there is no
risk of an attacker manipulating the encrypted file, an integrity
check seems a waste of resources.