fix: brighten/darken not reflecting any change
This bug was caused by an improper and wasteful check in
Helpers.lightDial method. It is improper because the floating-point
comparison is not accurate. It is wasteful since an improper input value
will end up throwing an exception and the UI code (responsible for calling
brighten/darken method) handles the input itself.